functions have moved namespaces fixing imports
