+++ rosetta-source/src/main/java/org/finos/cdm/CdmRuntimeModule.java
@@ -2,7 +2,7 @@ package org.finos.cdm;
 
 import cdm.base.datetime.functions.*;
 import cdm.base.math.functions.*;
-import cdm.observable.asset.fro.functions.IndexValueObservation;
+import cdm.observable.asset.functions.IndexValueObservation;
 import cdm.observable.asset.fro.functions.IndexValueObservationEmptyDataProvider;
 import cdm.product.collateral.functions.MergeEligibleCollateralCriteria;
 import cdm.product.collateral.functions.MergeEligibleCollateralCriteriaImpl;
