+++ rosetta-source/src/main/rosetta/product-asset-floatingrate-func.rosetta
@@ -6,6 +6,7 @@ import cdm.base.math.*
 
 import cdm.observable.asset.calculatedrate.*
 import cdm.observable.asset.fro.*
+import cdm.observable.asset.*
 
 import cdm.product.asset.*
 import cdm.product.asset.calculation.*
