+++ rosetta-source/src/test/java/cdm/product/asset/calculation/functions/FloatingAmountCalculationTest.java
@@ -3,9 +3,8 @@ package cdm.product.asset.calculation.functions;
 import cdm.base.datetime.*;
 import cdm.base.datetime.daycount.DayCountFractionEnum;
 import cdm.base.datetime.metafields.FieldWithMetaBusinessCenterEnum;
-import cdm.observable.asset.FloatingRateIndex;
 import cdm.observable.asset.InterestRateIndex;
-import cdm.observable.asset.fro.functions.IndexValueObservation;
+import cdm.observable.asset.functions.IndexValueObservation;
 import cdm.product.asset.InterestRatePayout;
 import cdm.product.asset.RateSpecification;
 import cdm.product.asset.floatingrate.FloatingAmountCalculationDetails;
