+++ rosetta-source/src/test/java/cdm/product/asset/floatingrate/functions/EvaluateScreenRateTest.java
@@ -1,9 +1,8 @@
 package cdm.product.asset.floatingrate.functions;
 
 import cdm.base.datetime.BusinessCenterEnum;
-import cdm.observable.asset.FloatingRateIndex;
 import cdm.observable.asset.InterestRateIndex;
-import cdm.observable.asset.fro.functions.IndexValueObservation;
+import cdm.observable.asset.functions.IndexValueObservation;
 import cdm.product.asset.FloatingRate;
 import cdm.product.asset.floatingrate.FloatingRateSettingDetails;
 import cdm.product.common.schedule.CalculationPeriodBase;
