+++ rosetta-source/src/test/java/cdm/product/asset/floatingrate/functions/FloatingRateTestHelper.java
@@ -3,12 +3,11 @@ package cdm.product.asset.floatingrate.functions;
 import cdm.base.datetime.*;
 import cdm.base.math.DatedValue;
 import cdm.base.staticdata.asset.rates.FloatingRateIndexEnum;
-import cdm.base.staticdata.asset.rates.metafields.FieldWithMetaFloatingRateIndexEnum;
 import cdm.observable.asset.FloatingRateIndex;
 import cdm.observable.asset.InterestRateIndex;
 import cdm.observable.asset.PriceSchedule;
-import cdm.observable.asset.fro.functions.IndexValueObservation;
 import cdm.observable.asset.fro.functions.IndexValueObservationTestDataProvider;
+import cdm.observable.asset.functions.IndexValueObservation;
 import cdm.product.asset.FloatingRateSpecification;
 import cdm.product.asset.SpreadSchedule;
 import cdm.product.common.schedule.RateSchedule;
